fix: prevent rerender loop
